Update ZenKit (previously "phoenix") (#555)
* Upgrade ZenKit (previously "phoenix")

+++ game/ui/gamemenu.cpp
@@ -1197,7 +1197,7 @@ void GameMenu::setPlayer(const Npc &pl) {
     return;
     }
 
-  set("MENU_ITEM_PLAYERGUILD", gilds->get_string(pl.guild()));
+  set("MENU_ITEM_PLAYERGUILD", gilds->get_string(uint16_t(pl.guild())));
 
   set("MENU_ITEM_LEVEL",       pl.level());
   set("MENU_ITEM_EXP",         pl.experience());
@@ -1216,8 +1216,8 @@ void GameMenu::setPlayer(const Npc &pl) {
 
   const bool g2        = Gothic::inst().version().game==2;
   const int  talentMax = g2 ? TALENT_MAX_G2 : TALENT_MAX_G1;
-  for(int i=0; i<talentMax; ++i) {
-    auto& str = tal->get_string(size_t(i));
+  for(uint16_t i=0; i<talentMax; ++i) {
+    auto& str = tal->get_string(i);
     if(str.empty())
       continue;
 
@@ -1225,7 +1225,7 @@ void GameMenu::setPlayer(const Npc &pl) {
     const int val = g2 ? pl.hitChance(Talent(i)) : pl.talentValue(Talent(i));
 
     set(string_frm("MENU_ITEM_TALENT_",i,"_TITLE"), str);
-    set(string_frm("MENU_ITEM_TALENT_",i,"_SKILL"), strEnum(talV->get_string(size_t(i)),sk,textBuf));
+    set(string_frm("MENU_ITEM_TALENT_",i,"_SKILL"), strEnum(talV->get_string(i),sk,textBuf));
     set(string_frm("MENU_ITEM_TALENT_",i),          string_frm(val,"%"));
     }
   }
